Child Pedestrians involved in Fatal Motor Vehicle Crashes, 2009-2023

Daily data on child pedestrians (aged 0-17 years) involved in a motor vehicle crash that resulted in a fatality.

Usage

farsinvolved

Format

## `farsinvolved` A data frame with 5,490 rows and 4 columns:

month

Month (character)

day

Day of the month (character)

year

Year (character)

n

Number of pedestrians

Source

National Highway Traffic Safety Administration (NHTSA) Motor Vehicle Crash Data Querying and Reporting

Details

Each row is a day of the year between January 1st 2009 and December 31st 2023. The `n` column is the number of pedestrians in the United States who were involved in a motor vehicle crash that day, where the event resulted in a fatality and where the pedestrian was aged between 0 and 17 years old. The person killed is not necessarily the pedestrian.

Author

Kieran Healy