General Social Survey data, 1972-2024

A dataset containing an extract from the General Social Survey. See http://gss.norc.org/Get-Documentation for full documentation of the variables. This data contains many of the same variables as `gss_sm`, but for all available years from 1972-2024.

Usage

gss_lon

Format

A data frame with 75,699 rows and 25 columns.

Source

National Opinion Research Center, http://gss.norc.org.

Details

year. GSS year for this respondent.

id. Respondent id number.

ballot. Ballot used for interview.

age. Age of respondent.

degree. R's highest degree.

race. Race of respondent.

sex. Respondent's sex.

siblings. Number of brothers and sisters (recoded from SIBS).

kids. Number of children (recoded from CHILDS).

bigregion. Region of interview (identical with REGION).

region. Region of interview.

income16. Total family income.

religion. R's religious preference (recoded from RELIGION)

marital. Marital status.

padeg. Father's highest degree.

madeg. Mother's highest degree.

partyid. Political party affiliation.

polviews. Think of self as liberal or conservative.

happy. General happiness.

partners_rc. How many sex partners r had in last year. (Recoded from PARTNERS)

grass. Should marijuana be made legal.

zodiac. Respondent's astrological sign.

wtssall. Person weight variable (1972-2018).

wtssps. Person weight variable (1972-2024).

vpsu. Sampling unit

vstrat. Stratification unit