round_df

Round numeric columns of a data frame or tibble

round_df(data, dig = 2)

Arguments

data
A data frame or tibble

dig
The number of digits to round to

Value

An object of the same class as `data`, with the numeric columns rounded off to `dig`

Details

Takes a data frame or tibble as input, rounds the numeric columns to the specified number of digits.

Examples

round_df(iris, 0)

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1              5           4            1           0     setosa
#> 2              5           3            1           0     setosa
#> 3              5           3            1           0     setosa
#> 4              5           3            2           0     setosa
#> 5              5           4            1           0     setosa
#> 6              5           4            2           0     setosa
#> 7              5           3            1           0     setosa
#> 8              5           3            2           0     setosa
#> 9              4           3            1           0     setosa
#> 10             5           3            2           0     setosa
#> 11             5           4            2           0     setosa
#> 12             5           3            2           0     setosa
#> 13             5           3            1           0     setosa
#> 14             4           3            1           0     setosa
#> 15             6           4            1           0     setosa
#> 16             6           4            2           0     setosa
#> 17             5           4            1           0     setosa
#> 18             5           4            1           0     setosa
#> 19             6           4            2           0     setosa
#> 20             5           4            2           0     setosa
#> 21             5           3            2           0     setosa
#> 22             5           4            2           0     setosa
#> 23             5           4            1           0     setosa
#> 24             5           3            2           0     setosa
#> 25             5           3            2           0     setosa
#> 26             5           3            2           0     setosa
#> 27             5           3            2           0     setosa
#> 28             5           4            2           0     setosa
#> 29             5           3            1           0     setosa
#> 30             5           3            2           0     setosa
#> 31             5           3            2           0     setosa
#> 32             5           3            2           0     setosa
#> 33             5           4            2           0     setosa
#> 34             6           4            1           0     setosa
#> 35             5           3            2           0     setosa
#> 36             5           3            1           0     setosa
#> 37             6           4            1           0     setosa
#> 38             5           4            1           0     setosa
#> 39             4           3            1           0     setosa
#> 40             5           3            2           0     setosa
#> 41             5           4            1           0     setosa
#> 42             4           2            1           0     setosa
#> 43             4           3            1           0     setosa
#> 44             5           4            2           1     setosa
#> 45             5           4            2           0     setosa
#> 46             5           3            1           0     setosa
#> 47             5           4            2           0     setosa
#> 48             5           3            1           0     setosa
#> 49             5           4            2           0     setosa
#> 50             5           3            1           0     setosa
#> 51             7           3            5           1 versicolor
#> 52             6           3            4           2 versicolor
#> 53             7           3            5           2 versicolor
#> 54             6           2            4           1 versicolor
#> 55             6           3            5           2 versicolor
#> 56             6           3            4           1 versicolor
#> 57             6           3            5           2 versicolor
#> 58             5           2            3           1 versicolor
#> 59             7           3            5           1 versicolor
#> 60             5           3            4           1 versicolor
#> 61             5           2            4           1 versicolor
#> 62             6           3            4           2 versicolor
#> 63             6           2            4           1 versicolor
#> 64             6           3            5           1 versicolor
#> 65             6           3            4           1 versicolor
#> 66             7           3            4           1 versicolor
#> 67             6           3            4           2 versicolor
#> 68             6           3            4           1 versicolor
#> 69             6           2            4           2 versicolor
#> 70             6           2            4           1 versicolor
#> 71             6           3            5           2 versicolor
#> 72             6           3            4           1 versicolor
#> 73             6           2            5           2 versicolor
#> 74             6           3            5           1 versicolor
#> 75             6           3            4           1 versicolor
#> 76             7           3            4           1 versicolor
#> 77             7           3            5           1 versicolor
#> 78             7           3            5           2 versicolor
#> 79             6           3            4           2 versicolor
#> 80             6           3            4           1 versicolor
#> 81             6           2            4           1 versicolor
#> 82             6           2            4           1 versicolor
#> 83             6           3            4           1 versicolor
#> 84             6           3            5           2 versicolor
#> 85             5           3            4           2 versicolor
#> 86             6           3            4           2 versicolor
#> 87             7           3            5           2 versicolor
#> 88             6           2            4           1 versicolor
#> 89             6           3            4           1 versicolor
#> 90             6           2            4           1 versicolor
#> 91             6           3            4           1 versicolor
#> 92             6           3            5           1 versicolor
#> 93             6           3            4           1 versicolor
#> 94             5           2            3           1 versicolor
#> 95             6           3            4           1 versicolor
#> 96             6           3            4           1 versicolor
#> 97             6           3            4           1 versicolor
#> 98             6           3            4           1 versicolor
#> 99             5           2            3           1 versicolor
#> 100            6           3            4           1 versicolor
#> 101            6           3            6           2  virginica
#> 102            6           3            5           2  virginica
#> 103            7           3            6           2  virginica
#> 104            6           3            6           2  virginica
#> 105            6           3            6           2  virginica
#> 106            8           3            7           2  virginica
#> 107            5           2            4           2  virginica
#> 108            7           3            6           2  virginica
#> 109            7           2            6           2  virginica
#> 110            7           4            6           2  virginica
#> 111            6           3            5           2  virginica
#> 112            6           3            5           2  virginica
#> 113            7           3            6           2  virginica
#> 114            6           2            5           2  virginica
#> 115            6           3            5           2  virginica
#> 116            6           3            5           2  virginica
#> 117            6           3            6           2  virginica
#> 118            8           4            7           2  virginica
#> 119            8           3            7           2  virginica
#> 120            6           2            5           2  virginica
#> 121            7           3            6           2  virginica
#> 122            6           3            5           2  virginica
#> 123            8           3            7           2  virginica
#> 124            6           3            5           2  virginica
#> 125            7           3            6           2  virginica
#> 126            7           3            6           2  virginica
#> 127            6           3            5           2  virginica
#> 128            6           3            5           2  virginica
#> 129            6           3            6           2  virginica
#> 130            7           3            6           2  virginica
#> 131            7           3            6           2  virginica
#> 132            8           4            6           2  virginica
#> 133            6           3            6           2  virginica
#> 134            6           3            5           2  virginica
#> 135            6           3            6           1  virginica
#> 136            8           3            6           2  virginica
#> 137            6           3            6           2  virginica
#> 138            6           3            6           2  virginica
#> 139            6           3            5           2  virginica
#> 140            7           3            5           2  virginica
#> 141            7           3            6           2  virginica
#> 142            7           3            5           2  virginica
#> 143            6           3            5           2  virginica
#> 144            7           3            6           2  virginica
#> 145            7           3            6           2  virginica
#> 146            7           3            5           2  virginica
#> 147            6           2            5           2  virginica
#> 148            6           3            5           2  virginica
#> 149            6           3            5           2  virginica
#> 150            6           3            5           2  virginica