titanic {socviz}	R Documentation

A table of survival rates from the Titanic

Description

A small table of survival rates from the Titanic, by sex

Usage

titanic

Format

"A data frame with four rows and four columns."

Source

Titanic data

[Package socviz version 0.5.0 Index]