center_df {socviz}	R Documentation

center_df

Description

Scale and/or center the numeric columns of a data frame or tibble

Usage

center_df(data, sc = FALSE, cen = TRUE)

Arguments

data
A data frame or tibble

sc
Scale the variables (default FALSE)

cen
Center the variables on their means (default TRUE)

Details

Takes a data frame or tibble as input and scales and/or centers the numeric columns. By default, centers but doesn't scale

Value

An object of the same class as 'data', with the numeric columns scaled or centered as requested

Author(s)

Kieran Healy

[Package socviz version 0.5.0 Index]