US Presidential Election vote shares

A dataset of US presidential elections from 1824 to 2016, with information on the winner, runner up, and various measures of vote share. Data for 2016 are provisional as of early December 2016. The variables are as follows:

data(elections_historic)

Format

A (tibble) data frame with 237 rows and 21 variables.

Source

https://en.wikipedia.org/wiki/List_of_United_States_presidential_elections_by_popular_vote_margin.

Details

election. Number of the election counting from the first US presidential election. 1824 is the 10th election.

year. Year.

winner. Full name of winner.

win.party. Party affiliation of winner.

ec.pect. Winner's share of electoral college vote. (Range is 0 to 1.)

popular.pct. Winner's share of popular vote. (Range is 0 to 1.)

popular.margin. Winner's point margin in the popular vote. Can be positive or negative.

votes. Total votes cast in the election.

margin. Winner's vote margin in the popular vote.

runner.up. Runner up candidate.

ru.part. Party affiliation of runner up candidate.

turnout.pct. Voter turnout as a proportion of eligible voters. (Rate is 0 to 1.)

winner.lname Last name of winner.

winner.label Winner's last name and election year.

ru.lastname. Runner up's last name.

ru.label. Runner up's last name and election year.

two.term. Is this a two term presidency? (TRUE/FALSE.) Note that F.D. Roosevelt was elected four times.

ec.votes. Electoral college votes cast for winner.

ec.denom. Total number of electoral college votes.