freq_tab

Generate an n-way frequency table

freq_tab(df, ...)

Arguments

df
tibble or data frame (implicit within pipline)

...
grouping, as with group_by()

Value

A tibble with the grouping variables, the n per group, and the proportion (prop) of each group, calculated with respect to the outermost grouping variable.

Details

tidyverse, pipeline, and dplyr-friendly frequency tables

Examples

not_run({
  mtcars %>% freq_tab(vs, gear, carb)
})