freq_tab {socviz}	R Documentation

freq_tab

Description

Generate a tidy n-way frequency table

Usage

freq_tab(df, ...)

Arguments

df
tibble or data frame (implicit within pipline)

...
grouping, as with group_by()

Details

tidyverse, pipeline, and dplyr-friendly frequency tables

Value

A tibble with the grouping variables, the N ('n') per group, and the proportion ('prop') of each group, calculated with respect to the outermost grouping variable.

Author(s)

Kieran Healy

Examples

## Not run: 
mtcars %>% freq_tab(vs, gear, carb)

## End(Not run)

[Package socviz version 0.5.0 Index]