gss_lon {socviz}	R Documentation

General Social Survey data, 1972-2016

Description

A dataset containing an extract from the General Social Survey. See http://gss.norc.org/Get-Documentation for full documentation of the variables. This data contains the same variables as 'gss_sm', but for all available years from 1972-2016.

Usage

gss_lon

Format

A data frame with 62,366 rows and 26 variables.

Details

year. gss year for this respondent.

id. respondent id number.

ballot. ballot used for interview.

age. age of respondent.

degree. Rs highest degree.

race. race of respondent.

sex. respondent's sex.

siblings. Number of brothers and sisters (recoded from SIBS).

kids. Number of children (recoded from CHILDS).

bigregion. region of interview (recoded from REGION).

income16. total family income.

religion. rs religious preference (recoded from RELIGION)

marital. marital status.

padeg. fathers highest degree.

madeg. mothers highest degree.

partyid. political party affiliation.

polviews. think of self as liberal or conservative.

happy. general happiness.

partners_rc. how many sex partners r had in last year. (Recoded from PARTNERS)

grass. should marijuana be made legal.

zodiac. respondents astrological sign.

pres12. R's stated vote in the 2012 Presidential election

wtssall. weight variable.

vpsu. Sampling unit

vstrat. Stratification unit

Source

National Opinion Research Center, http://gss.norc.org.

[Package socviz version 0.5.0 Index]