int_to_year

Convert an integer to a date.

int_to_year(x, month = "01", day = "01")

Arguments

x
An integer or vector integers.

month
The month to be added to the year. Months 1 to 9 should be given as 01, 02, etc.

day
The day to be added to the year. Days should be given as 01 or 02, etc, and not 1 or 2, etc.

Value

A vector of dates where the input integer forms the year component. The day and month components added will by default be today's day and month. For input, only years 0:9999 are accepted.

Examples

int_to_year(1960)

#> [1] "1960-01-01"

class(int_to_year(1960))

#> [1] "Date"

int_to_year(1960:1965)

#> [1] "1960-01-01" "1961-01-01" "1962-01-01" "1963-01-01" "1964-01-01"
#> [6] "1965-01-01"