organdata {socviz}	R Documentation

Organ donation in the OECD

Description

A dataset containing data on rates of organ donation for seventeen OECD countries between 1991 and 2002. The variables are as follows:

Usage

organdata

Format

A (tibble) data frame with 237 rows and 21 variables.

Details

country. Country name.

year. Year.

donors. Organ Donation rate per million population.

pop. Population in thousands.

pop_dens. Population density per square mile.

gdp. Gross Domestic Product in thousands of PPP dollars.

gdp_lag. Lagged Gross Domestic Product in thousands of PPP dollars.

health. Health spending, thousands of PPP dollars per capita.

health_lag Lagged health spending, thousands of PPP dollars per capita.

pubhealth. Public health spending as a percentage of total expenditure.

roads. Road accident fatalities per 100,000 population.

cerebvas. Cerebrovascular deaths per 100,000 population (rounded).

assault. Assault deaths per 100,000 population (rounded).

external. Deaths due to external causes per 100,000 population.

txp_pop. Transplant programs per million population.

world. Welfare state world (Esping Andersen.)

opt. Opt-in policy or Opt-out policy.

consent_law. Consent law, informed or presumed.

consent_practice. Consent practice, informed or presumed.

consistent. Law consistent with practice, yes or no.

ccode. Abbreviated country code.

Source

Macro-economic and spending data: OECD. Other data: Kieran Healy.

[Package socviz version 0.5.0 Index]