prefix_strip {socviz}	R Documentation

prefix_strip

Description

Strip a series of characters from the beginning of a character vector.

Usage

prefix_strip(var_string, prefixes, toTitle = TRUE, ...)

Arguments

var_string
A character vector, usually variable names

prefixes
A character vector, usually variable prefixes

toTitle
Convert results to Title Case? Defaults to TRUE.

...
Other arguments to 'gsub'

Details

Takes a character vector (usually vector of variable names from a summarized or tidied model object), along with a vector of character terms (usually the prefix of a dummy or categorical variable added by R when creating model terms) and strips the latter away from the former. Useful for quickly cleaning variable names for a plot.

Value

A character vector with 'prefixes' terms stripped from the beginning of 'var_name' terms.

Author(s)

Kieran Healy

[Package socviz version 0.5.0 Index]