tw_tab {socviz}	R Documentation

tw_tab

Description

Quickly make a two-way table of proportions (percentages)

Usage

tw_tab(x, y, margin = NULL, digs = 1, dnn = NULL, ...)

Arguments

x
Row variable

y
Column variable

margin
See 'prop.table'. Default is joint distribution (all cells sum to 100), 1 for row margins (rows sum to 1), 2 for column margins (columns sum to 1)

digs
Number of rigits to round percentages to. Defaults to 1.

dnn
See 'table'. the names to be given to the dimensions in the result (the dimnames names). Defaults to NULL for none.

...
Other arguments to be passed to 'table'.

Details

A wrapper for 'table' and 'prop.table' with the margin labels set by default to NULL and the cells rounded to percents at 1 decimal place.

Value

A contingency table of percentage values.

Author(s)

Kieran Healy

[Package socviz version 0.5.0 Index]